\name{annotatePeakInBatch.example2}
\alias{annotatePeakInBatch.example2}
\title{annotatePeakInBatch.example2}
\usage{
  annotatePeakInBatch.example2()
}
\description{
  you have a list of transcription factor biding sites from
  literature and are interested in determining the extent
  of the overlap to the list of peaks from your experiment.
  Prior calling the function annotatePeakInBatch, need to
  represent both dataset as RangedData where start is the
  start of the binding site, end is the end of the binding
  site, names is the name of the binding site, space and
  strand are the chromosome name and strand where the
  binding site is located.
}
\author{
  Julie Zhu
}

