\docType{data}
\name{CpG.island.data}
\alias{CpG}
\alias{CpG.human.GRCh37}
\alias{CpG.human.hg17}
\alias{CpG.human.hg18}
\alias{CpG.human.NCBI35}
\alias{CpG.human.NCBI36}
\alias{CpG.island.data}
\alias{CpG.mouse.mm8}
\alias{CpG.mouse.mm9}
\alias{CpG.mouse.NCBI36}
\alias{CpG.mouse.NCBI37}
\alias{CpG.rat.RGSC3.4}
\alias{CpG.rat.rn4}
\alias{pG.human.hg19}
\title{CpG islands}
\format{a \code{RangedData} instance, 1 row per CpG island.}
\source{
  UCSC Table Browser
}
\description{
  These datasets list the genomic locations of CpG islands
  from the named species, with coordinates based on the
  genome build. They are named: \code{Cpg.org.build}, eg
  \code{CpG.human.GRCh37}, or \code{CpG.mouse.NCBI37}
}
\author{
  Mark Cowley, 2012-03-05
}
\keyword{datasets}

