\docType{data}
\name{CpG.human.NCBI35}
\alias{CpG.human.hg17}
\alias{CpG.human.NCBI35}
\title{CpG islands - human - genome build: NCBI35/hg17}
\format{a \code{RangedData} instance, 1 row per CpG island.}
\source{
  UCSC Table Browser
}
\description{
  This data set list the genomic locations of human CpG
  islands, with coordinates based on the NCBI35 / hg17
  genome build.
}
\author{
  Mark Cowley, 2012-03-05
}
\keyword{datasets}

