\docType{data}
\name{CpG.human.NCBI36}
\alias{CpG.human.hg18}
\alias{CpG.human.NCBI36}
\title{CpG islands - human - genome build: NCBI36/hg18}
\format{a \code{RangedData} instance, 1 row per CpG island.}
\source{
  UCSC Table Browser
}
\description{
  This data set list the genomic locations of human CpG
  islands, with coordinates based on the NCBI36 / hg18
  genome build.
}
\author{
  Mark Cowley, 2012-03-05
}
\keyword{datasets}

