\docType{data}
\name{CpG.mouse.NCBI36}
\alias{CpG.mouse.mm8}
\alias{CpG.mouse.NCBI36}
\title{CpG islands - mouse - genome build: NCBI36/mm8}
\format{a \code{RangedData} instance, 1 row per CpG island.}
\source{
  UCSC Table Browser
}
\description{
  This data set list the genomic locations of mouse CpG
  islands, with coordinates based on the NCBI36 / mm8
  genome build.
}
\author{
  Mark Cowley, 2012-03-05
}
\keyword{datasets}

