\docType{data}
\name{CpG.mouse.NCBI37}
\alias{CpG.mouse.mm9}
\alias{CpG.mouse.NCBI37}
\title{CpG islands - mouse - genome build: NCBI37/mm9}
\format{a \code{RangedData} instance, 1 row per CpG island.}
\source{
  UCSC Table Browser
}
\description{
  This data set list the genomic locations of mouse CpG
  islands, with coordinates based on the NCBI37 / mm9
  genome build.
}
\author{
  Mark Cowley, 2012-03-05
}
\keyword{datasets}

