\docType{data}
\name{CpG.rat.RGSC3.4}
\alias{CpG.rat.RGSC3.4}
\alias{CpG.rat.rn4}
\title{CpG islands - rat - genome build: RGSC3.4/rn4}
\format{a \code{RangedData} instance, 1 row per CpG island.}
\source{
  UCSC Table Browser
}
\description{
  This data set list the genomic locations of rat CpG
  islands, with coordinates based on the RGSC3.4 / rn4
  genome build.
}
\author{
  Mark Cowley, 2012-03-05
}
\keyword{datasets}

