% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_MACS.R
\name{install_MACS}
\alias{install_MACS}
\title{install MACS python package}
\usage{
install_MACS(envname = "MACS", method = "conda")
}
\arguments{
\item{envname}{Virtual environment to install the MACS python
module}
}
\value{
install MACS python module
}
\description{
install MACS python package
}
