\name{is.dir}
\alias{is.dir}
\title{Is the given path a directory?}
\usage{
  is.dir(path)
}
\arguments{
  \item{path}{the path to a file or directory}
}
\value{
  \dQuote{TRUE} if the path is a directory, \dQuote{FALSE}
  otherwise
}
\description{
  Is the given path a directory?
}
\author{
  Mark Cowley
}

