\name{is.macs.fixed}
\alias{is.macs.fixed}
\title{has a MACS result dir been fixed already?}
\usage{
  is.macs.fixed(dir)
}
\arguments{
  \item{dir}{a character(1) pointing to a directory
  containing MACS results}
}
\value{
  logical
}
\description{
  has a MACS result dir been fixed already?
}
\author{
  Mark Cowley, 2012-03-08
}

