% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{macsList-class}
\alias{macsList-class}
\title{macsList}
\arguments{
\item{arguments}{The arguments used in the function.}

\item{outputs}{The outputs from the function.}

\item{log}{The run logs.}
}
\description{
macsList
}
