% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdgdiff.R
\name{bdgdiff}
\alias{bdgdiff}
\title{bdgdiff}
\usage{
bdgdiff(
  t1bdg,
  t2bdg,
  c1bdg,
  c2bdg,
  cutoff = 3,
  minlen = 200L,
  maxgap = 100L,
  depth1 = 1,
  depth2 = 1,
  outdir = ".",
  oprefix = character(),
  outputfile = list(),
  log = TRUE
)
}
\arguments{
\item{t1bdg}{MACS pileup bedGraph for condition 1. Incompatible
with callpeak --SPMR output. REQUIRED}

\item{t2bdg}{MACS pileup bedGraph for condition 2. Incompatible
with callpeak --SPMR output. REQUIRED}

\item{c1bdg}{MACS control lambda bedGraph for condition
1. Incompatible with callpeak --SPMR output. REQUIRED}

\item{c2bdg}{MACS control lambda bedGraph for condition
2. Incompatible with callpeak --SPMR output. REQUIRED}

\item{cutoff}{logLR cutoff. DEFAULT: 3 (likelihood
ratio=1000)", default = 3}

\item{minlen}{Minimum length of differential region. Try bigger value to remove small regions. DEFAULT: 200",
default = 200}

\item{maxgap}{Maximum gap to merge nearby differential
regions. Consider a wider gap for broad marks. Maximum gap
should be smaller than minimum length (-g). DEFAULT:
100", default = 100}

\item{depth1}{Sequencing depth (# of non-redundant reads in
million) for condition 1. It will be used together with
--d2. See description for --d2 below for how to assign
them. Default: 1}

\item{depth2}{Sequencing depth (# of non-redundant reads in
million) for condition 2. It will be used together with
--d1. DEPTH1 and DEPTH2 will be used to calculate scaling
factor for each sample, to down-scale larger sample to the
level of smaller one. For example, while comparing 10 million
condition 1 and 20 million condition 2, use --d1 10 --d2 20,
then pileup value in bedGraph for condition 2 will be divided
by 2. Default: 1}

\item{outdir}{The output directory.}

\item{oprefix}{Output file prefix. Actual files will be named as
PREFIX_cond1.bed, PREFIX_cond2.bed and
PREFIX_common.bed. Mutually exclusive with -o/--ofile.}

\item{outputfile}{Output filenames. Must give three arguments in
order: 1. file for unique regions in condition 1; 2. file for
unique regions in condition 2; 3. file for common regions in
both conditions. Note: mutually exclusive with --o-prefix.}

\item{log}{Whether to capture logs.}
}
\description{
Differential peak detection based on paired four bedgraph
files. Note: All regions on the same chromosome in the bedGraph
file should be continuous so only bedGraph files from MACS3 are
accpetable.
}
