% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdgpeakcall.R
\name{bdgpeakcall}
\alias{bdgpeakcall}
\title{bdgpeakcall}
\usage{
bdgpeakcall(
  ifile,
  cutoff = 5,
  minlen = 200L,
  maxgap = 30L,
  call_summits = FALSE,
  cutoff_analysis = FALSE,
  trackline = TRUE,
  outdir = ".",
  outputfile = character(),
  log = TRUE
)
}
\arguments{
\item{ifile}{MACS score in bedGraph. REQUIRED.}

\item{cutoff}{Cutoff depends on which method you used for score
track. If the file contains pvalue scores from MACS3, score 5
means pvalue 1e-5. DEFAULT: 5", default = 5.}

\item{minlen}{minimum length of peak, better to set it as d
value. DEFAULT: 200", default = 200.}

\item{maxgap}{maximum gap between significant points in a peak,
better to set it as tag size. DEFAULT: 30", default = 30.}

\item{call_summits}{If set, MACS will use a more sophisticated
approach to find all summits in each enriched peak region
DEFAULT: False",default=False.}

\item{cutoff_analysis}{While set, bdgpeakcall will analyze number
or total length of peaks that can be called by different cutoff
then output a summary table to help user decide a better
cutoff. Note, minlen and maxgap may affect the
results. DEFAULT: False", default = False.}

\item{trackline}{Tells MACS not to include trackline with bedGraph
files. The trackline is required by UCSC.}

\item{outdir}{The output directory.}

\item{outputfile}{The output file.}

\item{log}{Whether to capture logs.}
}
\description{
Call peaks from bedGraph output. Note: All regions on the same
chromosome in the bedGraph file should be continuous so only
bedGraph files from MACS3 are accpetable.
}
\examples{
\dontrun{
bdgpeakcall("run_bdgcmp_FE.bdg", cutoff = 2)
}
}
