% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdgpeakcall.R
\name{bdgpeakcall}
\alias{bdgpeakcall}
\title{bdgpeakcall}
\usage{
bdgpeakcall(
  ifile,
  cutoff = 5,
  minlen = 200L,
  maxgap = 30L,
  call_summits = FALSE,
  cutoff_analysis = FALSE,
  trackline = TRUE,
  outdir = ".",
  outputfile = character(),
  log = TRUE,
  verbose = 2L
)
}
\arguments{
\item{ifile}{MACS score in bedGraph. REQUIRED.}

\item{cutoff}{Cutoff depends on which method you used for score
track. If the file contains pvalue scores from MACS3, score 5
means pvalue 1e-5. DEFAULT: 5", default = 5.}

\item{minlen}{minimum length of peak, better to set it as d
value. DEFAULT: 200", default = 200.}

\item{maxgap}{maximum gap between significant points in a peak,
better to set it as tag size. DEFAULT: 30", default = 30.}

\item{call_summits}{If set, MACS will use a more sophisticated
approach to find all summits in each enriched peak region
DEFAULT: False",default=False.}

\item{cutoff_analysis}{While set, bdgpeakcall will analyze number
or total length of peaks that can be called by different cutoff
then output a summary table to help user decide a better
cutoff. Note, minlen and maxgap may affect the
results. DEFAULT: False", default = False.}

\item{trackline}{Tells MACS not to include trackline with bedGraph
files. The trackline is required by UCSC.}

\item{outdir}{The output directory.}

\item{outputfile}{The output file.}

\item{log}{Whether to capture logs.}

\item{verbose}{Set verbose level of runtime message. 0: only show
critical message, 1: show additional warning message, 2: show
process information, 3: show debug messages. DEFAULT:2}
}
\value{
\code{macsList} object.
}
\description{
Call peaks from bedGraph output. Note: All regions on the same
chromosome in the bedGraph file should be continuous so only
bedGraph files from MACS3 are accpetable.
}
\examples{
\donttest{
eh <- ExperimentHub::ExperimentHub()
CHIP <- eh[["EH4558"]]
CTRL <- eh[["EH4563"]]
p1 <- pileup(CHIP, outdir = tempdir(),
             outputfile = "pileup_ChIP_bed.bdg", format = "BED")
p2 <- pileup(CTRL, outdir = tempdir(),
             outputfile = "pileup_CTRL_bed.bdg", format = "BED")
c1 <- bdgcmp(p1$outputs, p2$outputs, outdir = tempdir(),
             oprefix = "bdgcmp", pseudocount = 1, method = "FE")
bdgpeakcall(c1$outputs, cutoff = 2,
            outdir = tempdir(), outputfile = "bdgpeakcall")
}
}
